+++ skued/__init__.py
@@ -44,6 +44,7 @@ from .eproperties import (
 from .image import (
     align,
     azimuthal_average,
+    detector_wavevectors,
     combine_masks,
     ialign,
     isnr,